created common function for timeZone
